// src/api/models/bannerCreate.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend
 * Coded by Meu TEAM
 * OpenAPI spec version: 1.0.0
 */

export interface BannerCreate {
  /**
   * @maxLength 255
   * @nullable
   */
  title?: string | null;
  /**
   * @maxLength 255
   * @nullable
   */
  title_en?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  description_en?: string | null;
  /** @nullable */
  image_id?: string | null;
  /** @nullable */
  display_order?: number | null;
  /** @nullable */
  is_active?: boolean | null;
}