// src/api/models/getApiV10BannerParams.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend
 * Coded by Meu TEAM
 * OpenAPI spec version: 1.0.0
 */
import type { FiltersParameter } from './filtersParameter';
import type { SortFieldParameter } from './sortFieldParameter';
import type { SortOrderParameter } from './sortOrderParameter';
import type { PageParameter } from './pageParameter';
import type { PageSizeParameter } from './pageSizeParameter';

export type GetApiV10BannerParams = {
/**
 * filter, visit https://www.npmjs.com/package/sequelize-api-paginate for syntax
 */
filters?: FiltersParameter;
/**
 * sortField, visit https://www.npmjs.com/package/sequelize-api-paginate for syntax
 */
sortField?: SortFieldParameter;
/**
 * sort order, visit https://www.npmjs.com/package/sequelize-api-paginate for syntax
 */
sortOrder?: SortOrderParameter;
/**
 * page, visit https://www.npmjs.com/package/sequelize-api-paginate for syntax
 * @minimum 1
 */
page?: PageParameter;
/**
 * pageSize, visit https://www.npmjs.com/package/sequelize-api-paginate for syntax
 * @minimum 1
 */
pageSize?: PageSizeParameter;
};